requirements.txt
# UI Libraries
streamlit==1.31.1
streamlit-option-menu==0.3.12

# YOLOv11 - Object Detection
ultralytics==8.1.20

# PaddleOCR - Text Detection (GPU Version - CUDA 11.8)
paddlepaddle-gpu==2.6.1
paddleocr==2.7.3

# TrOCR - Text Recognition (Hugging Face) - GPU Version
transformers==4.37.2
torch==2.2.0+cu118
torchvision==0.17.0+cu118
torchaudio==2.2.0+cu118

# Image Processing & Data Handling
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.24.3
pandas==2.2.0
